refactor: 클래스에 적합한 패키지를 구성한다
