+++ src/main/java/racingcar/common/ExceptionCode.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.common;
 
 public enum ExceptionCode {
