+++ src/main/java/racingcar/common/GameSettings.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.common;
 
 public record GameSettings(int randomMinValue, int randomMaxValue, int moveThreshold) {
 }
